Add files via upload
This is a Machine Learning Model involving Multiple Linear Regression for a US bike-sharing provider BoomBikes that has recently suffered considerable dips in their revenues due to the ongoing Corona pandemic. The company is finding it very difficult to sustain in the current market scenario. The goal is to come up with a mindful business plan to be able to accelerate its revenue as soon as the ongoing lock down comes to an end, and the economy restores to a healthy state.